+++ cli/src/main/java/com/github/olivergondza/dumpling/cli/GroovyCommand.java
@@ -43,7 +43,7 @@ import org.kohsuke.args4j.Argument;
 import org.kohsuke.args4j.CmdLineException;
 import org.kohsuke.args4j.Option;
 
-import com.github.olivergondza.dumpling.groovy.GroovyInterpretterConfig;
+import com.github.olivergondza.dumpling.groovy.InterpretterConfig;
 import com.github.olivergondza.dumpling.model.ModelObject;
 import com.github.olivergondza.dumpling.model.ModelObject.Mode;
 import com.github.olivergondza.dumpling.model.ProcessRuntime;
@@ -55,7 +55,7 @@ import com.github.olivergondza.dumpling.model.ProcessRuntime;
  */
 public class GroovyCommand implements CliCommand {
 
-    private static final GroovyInterpretterConfig CONFIG = new GroovyInterpretterConfig();
+    private static final InterpretterConfig CONFIG = new InterpretterConfig();
 
     @Option(name = "-i", aliases = {"--in"}, usage = "Input for process runtime")
     private ProcessRuntime<?, ?, ?> runtime;
