+++ cli/src/main/java/com/github/olivergondza/dumpling/query/DeadlocksCommand.java
@@ -33,7 +33,7 @@ import com.github.olivergondza.dumpling.cli.ProcessStream;
 import com.github.olivergondza.dumpling.model.ProcessRuntime;
 import com.github.olivergondza.dumpling.query.Deadlocks.Result;
 
-public final static class DeadlocksCommand implements CliCommand {
+public final class DeadlocksCommand implements CliCommand {
 
     @Option(name = "-i", aliases = {"--in"}, required = true, usage = "Input for process runtime")
     private ProcessRuntime<?, ?, ?> runtime;
