+++ cli/src/test/resources/com/github/olivergondza/dumpling/factory/ThreadDumpFactoryVendorTest/multipleMonitorsOnSingleStackFrame.groovy
