+++ core/src/test/java/com/github/olivergondza/dumpling/query/DeadlocksTest.java
@@ -25,6 +25,7 @@ package com.github.olivergondza.dumpling.query;
 
 import static org.hamcrest.CoreMatchers.equalTo;
 import static org.hamcrest.MatcherAssert.assertThat;
+import static org.hamcrest.Matchers.containsString;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
 
@@ -33,7 +34,6 @@ import java.util.Set;
 import org.junit.Test;
 
 import com.github.olivergondza.dumpling.Util;
-import com.github.olivergondza.dumpling.cli.AbstractCliTest;
 import com.github.olivergondza.dumpling.factory.JvmRuntimeFactory;
 import com.github.olivergondza.dumpling.factory.ThreadDumpFactory;
 import com.github.olivergondza.dumpling.model.dump.ThreadDumpRuntime;
@@ -44,7 +44,7 @@ import com.github.olivergondza.dumpling.model.jvm.JvmThread;
 import com.github.olivergondza.dumpling.model.jvm.JvmThreadSet;
 import com.github.olivergondza.dumpling.query.Deadlocks.Result;
 
-public class DeadlocksTest extends AbstractCliTest {
+public class DeadlocksTest {
 
     volatile boolean running = false;
     @Test
@@ -131,66 +131,6 @@ public class DeadlocksTest extends AbstractCliTest {
         assertThat(deadlock.size(), equalTo(2));
 
         assertThat("Involved thread count", result.involvedThreads().size(), equalTo(2));
-    }
-
-    @Test
-    public void cliQuery() throws Exception {
-        run("deadlocks", "--in", "threaddump", Util.resourceFile("deadlock.log").getAbsolutePath());
-        assertThat(err.toString(), equalTo(""));
-        assertListing(out.toString());
-
-        assertThat(exitValue, equalTo(1));
-    }
-
-    @Test
-    public void toStringNoTraces() throws Exception {
-        ThreadDumpRuntime runtime = new ThreadDumpFactory().fromFile(Util.resourceFile("deadlock.log"));
-        assertListing(new Deadlocks().query(runtime.getThreads()).toString());
-    }
-
-    private void assertListing(String out) {
-        assertThat(out, containsString(Util.multiline(
-                "Deadlock #1:",
-                "\"Handling POST /hudson/job/some_other_job/doRename : ajp-127.0.0.1-8009-24\" daemon prio=10 tid=1481750528 nid=27336",
-                "\tWaiting to <0x40dce6960> (a hudson.model.ListView)",
-                "\tAcquired   <0x40dce0d68> (a hudson.plugins.nested_view.NestedView)",
-                "\tAcquired   <0x49c5f7990> (a hudson.model.FreeStyleProject)",
-                "\tAcquired * <0x404325338> (a hudson.model.Hudson)",
-                "\"Handling POST /hudson/view/some_view/configSubmit : ajp-127.0.0.1-8009-107\" daemon prio=10 tid=47091108077568 nid=17982",
-                "\tWaiting to <0x404325338> (a hudson.model.Hudson)",
-                "\tAcquired * <0x40dce6960> (a hudson.model.ListView)"
-        )));
-        assertThat(out, containsString("%nDeadlocks: 1%n"));
-    }
-
-    @Test
-    public void cliQueryTraces() throws Exception {
-        run("deadlocks", "--show-stack-traces", "--in", "threaddump", Util.resourceFile("deadlock.log").getAbsolutePath());
-        assertThat(err.toString(), equalTo(""));
-        assertLongListing(out.toString());
-
-        assertThat(exitValue, equalTo(1));
-    }
-
-    @Test
-    public void toStringWithTraces() throws Exception {
-        ThreadDumpRuntime runtime = new ThreadDumpFactory().fromFile(Util.resourceFile("deadlock.log"));
-        assertLongListing(new Deadlocks().showStackTraces().query(runtime.getThreads()).toString());
-    }
-
-    private void assertLongListing(String out) {
-        assertThat(out, containsString(Util.multiline(
-                "Deadlock #1:",
-                "\"Handling POST /hudson/job/some_other_job/doRename : ajp-127.0.0.1-8009-24\" daemon prio=10 tid=1481750528 nid=27336",
-                "\tWaiting to <0x40dce6960> (a hudson.model.ListView)",
-                "\tAcquired   <0x40dce0d68> (a hudson.plugins.nested_view.NestedView)",
-                "\tAcquired   <0x49c5f7990> (a hudson.model.FreeStyleProject)",
-                "\tAcquired * <0x404325338> (a hudson.model.Hudson)",
-                "\"Handling POST /hudson/view/some_view/configSubmit : ajp-127.0.0.1-8009-107\" daemon prio=10 tid=47091108077568 nid=17982",
-                "\tWaiting to <0x404325338> (a hudson.model.Hudson)",
-                "\tAcquired * <0x40dce6960> (a hudson.model.ListView)"
-        )));
-        assertThat(out, containsString("%nDeadlocks: 1%n"));
     }
 
     @Test
