+++ groovy-api/pom.xml
@@ -16,15 +16,12 @@
       <artifactId>dumpling</artifactId>
       <version>${project.version}</version>
     </dependency>
+    <dependency>
+      <groupId>com.github.olivergondza</groupId>
+      <artifactId>dumpling-test-utils</artifactId>
+      <scope>test</scope>
+    </dependency>
 
-    <dependency>
-      <groupId>javax.annotation</groupId>
-      <artifactId>jsr250-api</artifactId>
-    </dependency>
-    <dependency>
-      <groupId>org.reflections</groupId>
-      <artifactId>reflections</artifactId>
-    </dependency>
     <dependency>
       <groupId>org.codehaus.groovy</groupId>
       <artifactId>groovy-jsr223</artifactId>
