+++ test-utils/pom.xml
@@ -5,24 +5,15 @@
     <artifactId>dumpling-parent</artifactId>
     <version>1.1-SNAPSHOT</version>
   </parent>
-  <artifactId>dumpling</artifactId>
+  <artifactId>dumpling-test-utils</artifactId>
   <packaging>jar</packaging>
-  <name>Dumpling DSL</name>
-  <description>Dumpling core module</description>
+  <name>Dumpling test utils</name>
 
   <dependencies>
     <dependency>
-      <groupId>com.sun</groupId>
-      <artifactId>tools</artifactId>
-      <version>1.6</version>
-      <scope>system</scope>
-      <systemPath>${java.home}/../lib/tools.jar</systemPath>
-    </dependency>
-
-    <dependency>
-      <groupId>com.github.olivergondza</groupId>
-      <artifactId>dumpling-test-utils</artifactId>
-      <scope>test</scope>
+      <groupId>junit</groupId>
+      <artifactId>junit</artifactId>
+      <version>4.11</version>
     </dependency>
   </dependencies>
 </project>
